+++ src/cmdClientConfigureTools.go
@@ -56,7 +56,7 @@ func (c *clientConfigureToolsCmd) Execute(args []string) error {
 		tnodes = append(tnodes, no)
 	}
 	// store IP on tools nodes
-	err = b.CopyFilesToCluster(c.ClientName.String(), []fileList{{filePath: "/opt/asbench-grafana.ip", fileContents: strings.NewReader(ip), fileSize: len(ip)}}, tnodes)
+	err = b.CopyFilesToCluster(c.ClientName.String(), []fileList{{filePath: "/opt/asbench-grafana.ip", fileContents: ip, fileSize: len(ip)}}, tnodes)
 	if err != nil {
 		return fmt.Errorf("could not upload file 1: %s", err)
 	}
@@ -81,7 +81,7 @@ func (c *clientConfigureToolsCmd) Execute(args []string) error {
 		}
 		// install promtail if not found
 		promScript := promTailScript(isArm)
-		err = b.CopyFilesToCluster(string(c.ClientName), []fileList{{filePath: "/opt/install-promtail.sh", fileContents: strings.NewReader(promScript), fileSize: len(promScript)}}, []int{tnode})
+		err = b.CopyFilesToCluster(string(c.ClientName), []fileList{{filePath: "/opt/install-promtail.sh", fileContents: promScript, fileSize: len(promScript)}}, []int{tnode})
 		if err != nil {
 			return fmt.Errorf("failed to install loki download script: %s", err)
 		}
@@ -92,7 +92,7 @@ func (c *clientConfigureToolsCmd) Execute(args []string) error {
 	}
 	// install promtail config file
 	promScript := promTailConf()
-	err = b.CopyFilesToCluster(string(c.ClientName), []fileList{{filePath: "/opt/configure-promtail.sh", fileContents: strings.NewReader(promScript), fileSize: len(promScript)}}, tnodes)
+	err = b.CopyFilesToCluster(string(c.ClientName), []fileList{{filePath: "/opt/configure-promtail.sh", fileContents: promScript, fileSize: len(promScript)}}, tnodes)
 	if err != nil {
 		return fmt.Errorf("failed to install conf script: %s", err)
 	}